+++ yggdrasil_decision_forests/port/python/ydf/learner/learner_test.py
@@ -317,7 +317,9 @@ class RandomForestLearnerTest(LearnerTest):
     vds_dataset = DatasetForTesting(vds_train, vds_test, pd_dataset.label)
 
     tuner = tuner_lib.RandomSearchTuner(
-        num_trials=5, automatic_search_space=True
+        num_trials=5,
+        automatic_search_space=True,
+        parallel_trials=2,
     )
     learner = specialized_learners.GradientBoostedTreesLearner(
         label=pd_dataset.label,
@@ -341,7 +343,9 @@ class RandomForestLearnerTest(LearnerTest):
     vds_dataset = DatasetForTesting(vds_train, vds_test, pd_dataset.label)
 
     tuner = tuner_lib.RandomSearchTuner(
-        num_trials=5, automatic_search_space=True
+        num_trials=5,
+        automatic_search_space=True,
+        parallel_trials=2,
     )
     learner = specialized_learners.GradientBoostedTreesLearner(
         label=pd_dataset.label,